solicitud arreglado

+++ app/Http/Controllers/AgregmiadoController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\agregmiado;
+use App\Models\sex;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -20,8 +21,8 @@ class AgregmiadoController extends Controller
                 'id_sexo' => 'required|string|max:255',
                 'NUP' => 'required|string|max:10|unique:agregmiados',
                 'NUE' => 'required|string|max:10|unique:agregmiados',
-                'RFC' => 'required|string|max:10|unique:agregmiados',
-                'NSS' => 'required|string|max:10|unique:agregmiados',
+                'RFC' => 'required|string|max:13|unique:agregmiados',
+                'NSS' => 'required|string|max:50|unique:agregmiados',
                 'fecha_nacimiento' => 'required|string|max:255',
                 'telefono' => 'required|string|max:10|unique:agregmiados',
                 'cuota' => 'required|',
@@ -54,6 +55,7 @@ class AgregmiadoController extends Controller
             $newUser = User::create([
                 'name' => $request->nombre,
                 'NUE'=> $request->NUE,
+                'id_agregmiado' => $newAgregmiado->id,
                 'password' => bcrypt($request->NUE),
                 'id_rol' => $request->id_rol
             ]);
@@ -76,7 +78,13 @@ class AgregmiadoController extends Controller
 
     public function getAgregmiado(){
         try{
-            $agregmiado = agregmiado::all();
+
+            $agregmiado = agregmiado::addSelect([
+                'genere' => sex::select('genere')
+                    ->whereColumn('id_sexo', 'id')
+            ])->get();
+
+            // $agregmiado = agregmiado::all();
             return response()->json([
                 'success' => true,
                 'message' => 'Lista de agregmiados',
@@ -91,9 +99,9 @@ class AgregmiadoController extends Controller
         }
     }
 
-    public function getAgregmiadoById($id){
+    public function getAgregmiadoByNUE($NUE){
         try{
-            $agregmiado = agregmiado::find($id);
+            $agregmiado = agregmiado::where('NUE', $NUE)->first();
             return response()->json([
                 'success' => true,
                 'message' => 'Agregmiado encontrado',
@@ -108,21 +116,32 @@ class AgregmiadoController extends Controller
         }
     }
 
-    public function updateAgregmiado(Request $request, $id){
+    public function updateAgregmiado(Request $request, $nue){
         try{
+            $agregmiado = agregmiado::where('NUE', $nue)->first();
+            $agremiadoUser = User::where('NUE', $nue)->first();
+
+            if(!$agregmiado){
+                return response()->json([
+                    'success' => false,
+                    'message' => 'No se encontro el agregmiado',
+                    'errors' => null
+                ], 404);
+            }
+
             $validateUser = Validator::make($request->all(),[
-                'nombre' => 'required|string|max:255',
-                'apellido_p' => 'required|string|max:255',
-                'apellido_m' => 'required|string|max:255',
-                'id_sexo' => 'required|string|max:255',
-                'NUP' => 'required|string|max:10|unique:agregmiados',
-                'NUE' => 'required|string|max:10|unique:agregmiados',
-                'RFC' => 'required|string|max:10|unique:agregmiados',
-                'NSS' => 'required|string|max:10|unique:agregmiados',
-                'fecha_nacimiento' => 'required|string|max:255',
-                'telefono' => 'required|string|max:10|unique:agregmiados',
-                'cuota' => 'required|',
-                'id_rol' => 'required|'
+                'nombre' => 'string|max:255',
+                'apellido_p' => 'string|max:255',
+                'apellido_m' => 'string|max:255',
+                'id_sexo' => 'integer',
+                'NUP' => 'string|max:10',
+                'NUE' => 'string|max:10',
+                'RFC' => 'string|max:13',
+                'NSS' => 'string|max:50',
+                'fecha_nacimiento' => 'string|max:255',
+                // 'telefono' => ,
+                'cuota' => 'integer',
+                'id_rol' => 'integer'
             ]);
 
             if($validateUser->fails()){
@@ -133,25 +152,36 @@ class AgregmiadoController extends Controller
                 ], 422);
             }
 
-            $agregmiado = agregmiado::find($id);
-            $agregmiado->nombre = $request->get('nombre');
-            $agregmiado->apellido_p = $request->get('apellido_p');
-            $agregmiado->apellido_m = $request->get('apellido_m');
-            $agregmiado->id_sexo = $request->get('id_sexo');
-            $agregmiado->NUP = $request->get('NUP');
-            $agregmiado->NUE = $request->get('NUE');
-            $agregmiado->RFC = $request->get('RFC');
-            $agregmiado->NSS = $request->get('NSS');
-            $agregmiado->fecha_nacimiento = $request->get('fecha_nacimiento');
-            $agregmiado->telefono = $request->get('telefono');
-            $agregmiado->cuota = $request->get('cuota');
-            $agregmiado->id_rol = $request->get('id_rol');
-            $agregmiado->save();
+            $agregmiado->update([
+                'nombre' => $request->get('nombre'),
+                'apellido_p' => $request->get('apellido_p'),
+                'apellido_m' => $request->get('apellido_m'),
+                'id_sexo' => $request->get('id_sexo'),
+                'NUP' => $request->get('NUP'),
+                'NUE' => $request->get('NUE'),
+                'RFC' => $request->get('RFC'),
+                'NSS' => $request->get('NSS'),
+                'fecha_nacimiento' => $request->get('fecha_nacimiento'),
+                'telefono' => $request->get('telefono'),
+                'cuota' => $request->get('cuota'),
+                'id_rol' => $request->get('id_rol')
+            ]);
+
+            $agremiadoUser->update([
+                'name' => $request->nombre,
+                'NUE'=> $request->NUE,
+                'id_agregmiado' => $agregmiado->id,
+                'password' => bcrypt($request->NUE),
+                'id_rol' => $request->id_rol
+            ]);
 
             return response()->json([
                 'success' => true,
                 'message' => 'Agregmiado actualizado correctamente',
-                'data' => $agregmiado
+                'data' => [
+                    'agregmiado' => $agregmiado,
+                    'user' => $agremiadoUser
+                ]
             ], 200);
 
         } catch (\Throwable $th){
@@ -167,12 +197,28 @@ class AgregmiadoController extends Controller
     public function deleteAgregmiado($id){
         try{
             $agregmiado = agregmiado::find($id);
+            $agremiadoUser = User::where('id_agregmiado', $id)->first();
+
+            if(!$agregmiado){
+                return response()->json([
+                    'success' => false,
+                    'message' => 'No se encontro el agregmiado',
+                    'errors' => null
+                ], 404);
+            }
+
+            $agremiadoUser->delete();
             $agregmiado->delete();
+
             return response()->json([
                 'success' => true,
                 'message' => 'Agregmiado eliminado correctamente',
-                'data' => $agregmiado
+                'data' => [
+                    'agregmiado' => $agregmiado,
+                    'user' => $agremiadoUser
+                ]
             ], 200);
+
         } catch (\Throwable $th){
             return response()->json([
                 'success' => false,
